requirements/common.txt
python-dotenv==1.0.1
loguru==0.7.2
ocpp==0.26.0
websockets==12.0
httpx==0.26.0
arrow==1.3.0
aio-pika==9.4.0
pydantic==2.5.3